RabbitVNC: modify input password
When using the "Edit and Connect" feature in the connection list, if you do not choose to "save password", the Enter Password dialog box appears again when you connect after you enter the password. The reason is that after the settings dialog, save the configuration file, and then use the configuration file to connect. At this time, if you do not save the password, the password will not be saved to the configuration file, so the Enter Password dialog will appear when connecting.