Fix macOS paths
The program does not start on macOS because it's unable to find it's own libraries (DYLD_LIBRARY_PATH) and Qt libraries (DYLD_FRAMEWORK_PATH). Also, the macOS quarantines the program libraries, so it recursively removes the quarantine attributes of the program files for the program to be able to run.